Print Objects of the S3 Class permutedReserve

Function to organize and print the output provided by the permutation bootstrap method implemented in the function permuteReserve()

Usage

# S3 method for class 'permutedReserve'
print(x, ...)

Arguments

x

an object of the class permutedReserve resulting from a call of the functions permuteReserve()

...

further arguments passed to print

Value

Displays information about the estimated reserve (by one of the estimation algorithms – PARALLAX, REACT, or MACRAME) and the overall reserve distribution resulting from a call of the permuteReserve() function

See also

Examples

## reserve point prediction by the PARALLAX method
output <- parallelReserve(CameronMutual)

## reserve distribution prediction by the permutation bootstrap
x <- permuteReserve(output, B = 100)
#>   |                                           | 0 % ~calculating    |>                                          | 1 % ~00s            |>                                          | 2 % ~00s            |>>                                         | 3 % ~00s            |>>                                         | 4 % ~00s            |>>>                                        | 5 % ~00s            |>>>                                        | 6 % ~00s            |>>>>                                       | 7 % ~00s            |>>>>                                       | 8 % ~00s            |>>>>                                       | 9 % ~00s            |>>>>>                                      | 10% ~00s            |>>>>>                                      | 11% ~00s            |>>>>>>                                     | 12% ~00s            |>>>>>>                                     | 13% ~00s            |>>>>>>>                                    | 14% ~00s            |>>>>>>>                                    | 15% ~00s            |>>>>>>>                                    | 16% ~00s            |>>>>>>>>                                   | 17% ~00s            |>>>>>>>>                                   | 18% ~00s            |>>>>>>>>>                                  | 19% ~00s            |>>>>>>>>>                                  | 20% ~00s            |>>>>>>>>>>                                 | 21% ~00s            |>>>>>>>>>>                                 | 22% ~00s            |>>>>>>>>>>                                 | 23% ~00s            |>>>>>>>>>>>                                | 24% ~00s            |>>>>>>>>>>>                                | 25% ~00s            |>>>>>>>>>>>>                               | 26% ~00s            |>>>>>>>>>>>>                               | 27% ~00s            |>>>>>>>>>>>>>                              | 28% ~00s            |>>>>>>>>>>>>>                              | 29% ~00s            |>>>>>>>>>>>>>                              | 30% ~00s            |>>>>>>>>>>>>>>                             | 31% ~00s            |>>>>>>>>>>>>>>                             | 32% ~00s            |>>>>>>>>>>>>>>>                            | 33% ~00s            |>>>>>>>>>>>>>>>                            | 34% ~00s            |>>>>>>>>>>>>>>>>                           | 35% ~00s            |>>>>>>>>>>>>>>>>                           | 36% ~00s            |>>>>>>>>>>>>>>>>                           | 37% ~00s            |>>>>>>>>>>>>>>>>>                          | 38% ~00s            |>>>>>>>>>>>>>>>>>                          | 39% ~00s            |>>>>>>>>>>>>>>>>>>                         | 40% ~00s            |>>>>>>>>>>>>>>>>>>                         | 41% ~00s            |>>>>>>>>>>>>>>>>>>>                        | 42% ~00s            |>>>>>>>>>>>>>>>>>>>                        | 43% ~00s            |>>>>>>>>>>>>>>>>>>>                        | 44% ~00s            |>>>>>>>>>>>>>>>>>>>>                       | 45% ~00s            |>>>>>>>>>>>>>>>>>>>>                       | 46% ~00s            |>>>>>>>>>>>>>>>>>>>>>                      | 47% ~00s            |>>>>>>>>>>>>>>>>>>>>>                      | 48% ~00s            |>>>>>>>>>>>>>>>>>>>>>>                     | 49% ~00s            |>>>>>>>>>>>>>>>>>>>>>>                     | 50% ~00s            |>>>>>>>>>>>>>>>>>>>>>>                     | 51% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>                    | 52% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>                    | 53% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>                   | 54% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>                   | 55% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>                  | 56% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>                  | 57% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>                  | 58% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>                 | 59% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>                 | 60% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>                | 61% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>                | 62% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>               | 63% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>               | 64% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>               | 65% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>              | 66% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>              | 67% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>             | 68% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>             | 69% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>            | 70% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>            | 71% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>            | 72% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>           | 73% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>           | 74% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>          | 75% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>          | 76% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>         | 77% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>         | 78% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>         | 79% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>        | 80% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>        | 81% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>       | 82% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>       | 83% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>      | 84% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>      | 85% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>      | 86% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>     | 87% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>     | 88% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>    | 89% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>    | 90% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>   | 91% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>   | 92% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>   | 93% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>  | 94% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>  | 95% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> | 96% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> | 97% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>| 98% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>| 99% ~00s            |>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>| 100% elapsed=00s  

## summary of the results
print(x) 
#> PARALLAX based reserving
#>     Est.Reserve    Est.Ultimate     Paid Amount    True Reserve 
#>            8540          113699          105159            7963 
#> Permutation bootstrap (B = 100)
#>      Boot.Mean        Std.Er.       BootCov%    BootVar.995 
#>       9011.124        871.512          9.672          1.238