Set Custom Color Styles for profileLadder Output

Function to set user-modified color layout for the run-off triangle visualization and the overall output presentation

Usage

set.fancy.print(
  color.known = "#333333",
  color.predicted = "#CC00CC",
  color.unknown = "#999999",
  color.info = "#CC00CC"
)

Arguments

color.known

Color (e.g., a hexadecimal code) for the run-off triangle part (the upper-left triangle)

color.predicted

Color (e.g., a hexadecimal code) for the predicted part of the run-off triangle (the bottom-right triangle)

color.unknown

Color (e.g., a hexadecimal code) for the 'unknown' future (the bottom-right triangle which is typically not avalaialble for insurance practice but is often provided for retrospective evaluations)

color.info

Color (e.g., a hexadecimal code) for the information messages in the outputs of the prediction functions parallelReserve(), mcReserve(), and permuteReserve()

Value

Sets the user-defined option for fancy print color styles

See also

Examples

## fancy print option for the run-off triangle 
print(as.profileLadder(observed(CameronMutual)), fancy.print = TRUE)
#> Run-off Reserving 
#>     Estimated Reserve    Estimated Ultimate           Paid Amount 
#>                    NA                    NA                105159 
#>          True Reserve 
#>                    NA 
#> 
#> Run-off triangle
#> 5244 	 9228 	10823 	11352 	11791 	12082 	12120 	12199 	12215 	12215 	
#> 5984 	 9939 	11725 	12346 	12746 	12909 	13034 	13109 	13113 	    . 	
#> 7452 	12421 	14171 	14752 	15066 	15354 	15637 	15720 	    . 	    . 	
#> 7115 	11117 	12488 	13274 	13662 	13859 	13872 	    . 	    . 	    . 	
#> 5753 	 8969 	 9917 	10697 	11135 	11282 	    . 	    . 	    . 	    . 	
#> 3937 	 6524 	 7989 	 8543 	 8757 	    . 	    . 	    . 	    . 	    . 	
#> 5127 	 8212 	 8976 	 9325 	    . 	    . 	    . 	    . 	    . 	    . 	
#> 5046 	 8006 	 8984 	    . 	    . 	    . 	    . 	    . 	    . 	    . 	
#> 5129 	 8202 	    . 	    . 	    . 	    . 	    . 	    . 	    . 	    . 	
#> 3689 	    . 	    . 	    . 	    . 	    . 	    . 	    . 	    . 	    . 	

## standard print option for the run-off triangle
print(as.profileLadder(observed(CameronMutual)), fancy.print = FALSE)
#> Run-off Reserving 
#>     Estimated Reserve    Estimated Ultimate           Paid Amount 
#>                    NA                    NA                105159 
#>          True Reserve 
#>                    NA 
#> 
#> Run-off triangle
#>       dev
#> origin    1     2     3     4     5     6     7     8     9    10
#>     1  5244  9228 10823 11352 11791 12082 12120 12199 12215 12215
#>     2  5984  9939 11725 12346 12746 12909 13034 13109 13113    NA
#>     3  7452 12421 14171 14752 15066 15354 15637 15720    NA    NA
#>     4  7115 11117 12488 13274 13662 13859 13872    NA    NA    NA
#>     5  5753  8969  9917 10697 11135 11282    NA    NA    NA    NA
#>     6  3937  6524  7989  8543  8757    NA    NA    NA    NA    NA
#>     7  5127  8212  8976  9325    NA    NA    NA    NA    NA    NA
#>     8  5046  8006  8984    NA    NA    NA    NA    NA    NA    NA
#>     9  5129  8202    NA    NA    NA    NA    NA    NA    NA    NA
#>     10 3689    NA    NA    NA    NA    NA    NA    NA    NA    NA

## PARALLAX based run-off triangle completion (fancy print)
options(profileLadder.fancy = TRUE)
parallelReserve(CameronMutual)
#> PARALLAX Reserving 
#>     Estimated Reserve    Estimated Ultimate           Paid Amount 
#>                  8540                113699                105159 
#>          True Reserve 
#>                  7963 
#> 
#> PARALLAX method (functional profile completion)
#> 5244 	 9228 	10823 	11352 	11791 	12082 	12120 	12199 	12215 	12215 	
#> 5984 	 9939 	11725 	12346 	12746 	12909 	13034 	13109 	13113 	13113 	
#> 7452 	12421 	14171 	14752 	15066 	15354 	15637 	15720 	15724 	15724 	
#> 7115 	11117 	12488 	13274 	13662 	13859 	13872 	13947 	13951 	13951 	
#> 5753 	 8969 	 9917 	10697 	11135 	11282 	11320 	11399 	11415 	11415 	
#> 3937 	 6524 	 7989 	 8543 	 8757 	 8904 	 8942 	 9021 	 9037 	 9037 	
#> 5127 	 8212 	 8976 	 9325 	 9539 	 9686 	 9724 	 9803 	 9819 	 9819 	
#> 5046 	 8006 	 8984 	 9333 	 9547 	 9694 	 9732 	 9811 	 9827 	 9827 	
#> 5129 	 8202 	 8966 	 9315 	 9529 	 9676 	 9714 	 9793 	 9809 	 9809 	
#> 3689 	 6276 	 7741 	 8295 	 8509 	 8656 	 8694 	 8773 	 8789 	 8789 	

## PARALLAX based run-off triangle completion (standard print)
options(profileLadder.fancy = FALSE)
parallelReserve(CameronMutual)
#> PARALLAX Reserving 
#>     Estimated Reserve    Estimated Ultimate           Paid Amount 
#>                  8540                113699                105159 
#>          True Reserve 
#>                  7963 
#> 
#> PARALLAX method (functional profile completion)
#>       dev
#> origin    1     2     3     4     5     6     7     8     9    10
#>     1  5244  9228 10823 11352 11791 12082 12120 12199 12215 12215
#>     2  5984  9939 11725 12346 12746 12909 13034 13109 13113 13113
#>     3  7452 12421 14171 14752 15066 15354 15637 15720 15724 15724
#>     4  7115 11117 12488 13274 13662 13859 13872 13947 13951 13951
#>     5  5753  8969  9917 10697 11135 11282 11320 11399 11415 11415
#>     6  3937  6524  7989  8543  8757  8904  8942  9021  9037  9037
#>     7  5127  8212  8976  9325  9539  9686  9724  9803  9819  9819
#>     8  5046  8006  8984  9333  9547  9694  9732  9811  9827  9827
#>     9  5129  8202  8966  9315  9529  9676  9714  9793  9809  9809
#>     10 3689  6276  7741  8295  8509  8656  8694  8773  8789  8789